requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.1 